+++ packages/devtools_extensions/lib/src/template/_simulated_devtools_environment/_simulated_devtools_controller.dart
@@ -117,22 +117,12 @@ class SimulatedDevToolsController extends DisposableController
 
   Future<void> hotReloadConnectedApp() async {
     await serviceManager.performHotReload();
-    messageLogs.add(
-      MessageLogEntry(
-        source: MessageSource.info,
-        message: 'Hot reload performed on connected app',
-      ),
-    );
+    logInfoEvent('Hot reload performed on connected app');
   }
 
   Future<void> hotRestartConnectedApp() async {
     await serviceManager.performHotRestart();
-    messageLogs.add(
-      MessageLogEntry(
-        source: MessageSource.info,
-        message: 'Hot restart performed on connected app',
-      ),
-    );
+    logInfoEvent('Hot restart performed on connected app');
   }
 
   void toggleTheme() {
@@ -141,6 +131,12 @@ class SimulatedDevToolsController extends DisposableController
       theme: darkThemeEnabled
           ? ExtensionEventParameters.themeValueLight
           : ExtensionEventParameters.themeValueDark,
+    );
+  }
+
+  void logInfoEvent(String message) {
+    messageLogs.add(
+      MessageLogEntry(source: MessageSource.info, message: message),
     );
   }