+++ packages/devtools_extensions/lib/src/template/_simulated_devtools_environment/_simulated_devtools_environment.dart
@@ -10,15 +10,18 @@ import 'package:devtools_app_shared/service.dart';
 import 'package:devtools_app_shared/ui.dart';
 import 'package:devtools_app_shared/utils.dart';
 import 'package:flutter/material.dart';
-import 'package:web/web.dart' hide Text;
+import 'package:flutter/services.dart';
+import 'package:web/web.dart' hide Text, Clipboard;
 
 import '../../api/api.dart';
 import '../../api/model.dart';
 import '../../utils.dart';
 import '../devtools_extension.dart';
 
-part '_connect_ui.dart';
 part '_simulated_devtools_controller.dart';
+part 'connection_ui/_connect.dart';
+part 'connection_ui/_dtd_connect.dart';
+part 'connection_ui/_vm_service_connect.dart';
 
 /// Wraps [child] in a simulated DevTools environment.
 ///
@@ -33,11 +36,14 @@ class SimulatedDevToolsWrapper extends StatefulWidget {
     super.key,
     required this.child,
     required this.requiresRunningApplication,
+    required this.onDtdConnectionChange,
   });
 
   final Widget child;
 
   final bool requiresRunningApplication;
+
+  final Future<void> Function(String?) onDtdConnectionChange;
 
   @override
   State<SimulatedDevToolsWrapper> createState() =>
@@ -51,9 +57,10 @@ class SimulatedDevToolsWrapperState extends State<SimulatedDevToolsWrapper>
 
   late final ScrollController scrollController;
 
-  late ConnectedState connectionState;
+  bool get vmServiceConnected => vmServiceConnectionState.connected;
+  late ConnectedState vmServiceConnectionState;
 
-  bool get connected => connectionState.connected;
+  bool dtdConnected = false;
 
   @override
   void initState() {
@@ -62,10 +69,17 @@ class SimulatedDevToolsWrapperState extends State<SimulatedDevToolsWrapper>
 
     scrollController = ScrollController();
 
-    connectionState = serviceManager.connectedState.value;
+    vmServiceConnectionState = serviceManager.connectedState.value;
     addAutoDisposeListener(serviceManager.connectedState, () {
       setState(() {
-        connectionState = serviceManager.connectedState.value;
+        vmServiceConnectionState = serviceManager.connectedState.value;
+      });
+    });
+
+    dtdConnected = dtdManager.hasConnection;
+    addAutoDisposeListener(dtdManager.connection, () {
+      setState(() {
+        dtdConnected = dtdManager.hasConnection;
       });
     });
   }
@@ -84,7 +98,7 @@ class SimulatedDevToolsWrapperState extends State<SimulatedDevToolsWrapper>
       builder: (context, constraints) {
         final availableWidth = constraints.maxWidth;
         const environmentPanelMinWidth =
-            VmServiceConnection.totalControlsWidth + 2 * defaultSpacing;
+            VmServiceConnectionDisplay.totalControlsWidth + 2 * defaultSpacing;
 
         final environmentPanelFraction =
             environmentPanelMinWidth / availableWidth;
@@ -129,49 +143,25 @@ class SimulatedDevToolsWrapperState extends State<SimulatedDevToolsWrapper>
                                 style: theme.textTheme.titleMedium,
                               ),
                               const PaddedDivider(),
-                              VmServiceConnection(
-                                connected: connected,
+                              VmServiceConnectionDisplay(
+                                connected: vmServiceConnected,
                                 simController: simController,
                               ),
-                              const SizedBox(height: denseSpacing),
-                              // TODO(kenz): figure out how to simulate the DTD
-                              // connection. This may be non-trivial.
+                              const PaddedDivider(),
+                              DTDConnectionDisplay(
+                                simController: simController,
+                                connected: dtdConnected,
+                                onConnectionChange: widget.onDtdConnectionChange,
+                              ),
                               _SimulatedApi(
                                 simController: simController,
                                 requiresRunningApplication:
                                     widget.requiresRunningApplication,
-                                connectedToApplication: connected,
+                                connectedToApplication: vmServiceConnected,
                               ),
                               const PaddedDivider(),
                               Expanded(
-                                child: Column(
-                                  crossAxisAlignment: CrossAxisAlignment.start,
-                                  children: [
-                                    Row(
-                                      mainAxisAlignment:
-                                          MainAxisAlignment.spaceBetween,
-                                      children: [
-                                        Text(
-                                          'Logs:',
-                                          style: theme.textTheme.titleMedium,
-                                        ),
-                                        DevToolsButton.iconOnly(
-                                          icon: Icons.clear,
-                                          outlined: false,
-                                          tooltip: 'Clear logs',
-                                          onPressed: () =>
-                                              simController.messageLogs.clear(),
-                                        ),
-                                      ],
-                                    ),
-                                    const PaddedDivider.thin(),
-                                    Expanded(
-                                      child: _LogMessages(
-                                        simController: simController,
-                                      ),
-                                    ),
-                                  ],
-                                ),
+                                child: _LogsView(simController: simController),
                               ),
                             ],
                           ),
@@ -250,6 +240,42 @@ class _SimulatedApi extends StatelessWidget {
             ),
         ],
       ),
+    );
+  }
+}
+
+class _LogsView extends StatelessWidget {
+  const _LogsView({required this.simController});
+
+  final SimulatedDevToolsController simController;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Text(
+              'Logs:',
+              style: Theme.of(context).textTheme.titleMedium,
+            ),
+            DevToolsButton.iconOnly(
+              icon: Icons.clear,
+              outlined: false,
+              tooltip: 'Clear logs',
+              onPressed: () => simController.messageLogs.clear(),
+            ),
+          ],
+        ),
+        const PaddedDivider.thin(),
+        Expanded(
+          child: _LogMessages(
+            simController: simController,
+          ),
+        ),
+      ],
     );
   }
 }