+++ packages/devtools_app_shared/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.10-wip
 * Add `DTDManager` class and export from `service.dart`.
+* Add `showDevToolsDialog` helper method.
 
 ## 0.0.9
 * Bump `package:web` to `^0.4.1`.