+++ packages/devtools_extensions/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.0.14-wip
 * Add a global `dtdManager` for interacting with the Dart Tooling Daemon.
+* Add support for connecting to the Dart Tooling Daemon from the 
+simulated DevTools environment.
+* Add help buttons to the VM Service and DTD connection text fields in the
+simulated DevTools environment.
 
 ## 0.0.13
 * Bump `package:web` to `^0.4.1`.